requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
python-dotenv==1.0.1
cryptography>=3.4.8
Authlib==1.3.0
gunicorn==21.2.0
